# api/docs/swagger.yaml
definitions:
...
  api.ErrResponse:
    properties:
      details:
        items:
          type: string
        type: array
      error:
        type: string
    type: object
...
  api.createProfileBody:
    properties:
      name:
        type: string
    type: object
...
  model.Account:
    properties:
      auth_methods:
        items:
          type: string
        type: array
      email:
        type: string
      email_verified:
        type: boolean
      error_reports_consent:
        type: boolean
      id:
        type: string
      mfa:
        $ref: '#/definitions/model.MFASettings'
      profiles:
        items:
          type: string
        type: array
      queries:
        type: integer
    type: object
...
  model.Advanced:
    properties:
      recursor:
        type: string
    required:
    - recursor
    type: object
...
  model.CustomRule:
    properties:
      action:
        type: string
      id:
        type: string
      value:
        type: string
    required:
    - action
    - id
    - value
    type: object
...
  model.DNSSECSettings:
    properties:
      enabled:
        type: boolean
      send_do_bit:
        type: boolean
    required:
    - enabled
    - send_do_bit
    type: object
  model.LogsSettings:
    properties:
      enabled:
        type: boolean
      log_clients_ips:
        type: boolean
      log_domains:
        type: boolean
      retention:
        $ref: '#/definitions/model.Retention'
    required:
    - enabled
    - log_clients_ips
    - log_domains
    - retention
    type: object
  model.MFASettings:
    properties:
      totp:
        $ref: '#/definitions/model.TotpSettings'
    type: object
  model.Privacy:
    properties:
      blocklists:
        items:
          type: string
        type: array
      default_rule:
        enum:
        - block
        - allow
        type: string
      subdomains_rule:
        enum:
        - block
        - allow
        type: string
    required:
    - default_rule
    - subdomains_rule
    type: object
  model.Profile:
    properties:
      account_id:
        type: string
      id:
        type: string
      name:
        type: string
      profile_id:
        type: string
      settings:
        $ref: '#/definitions/model.ProfileSettings'
    required:
    - account_id
    - id
    - name
    - profile_id
    - settings
    type: object
  model.ProfileSettings:
    properties:
      advanced:
        $ref: '#/definitions/model.Advanced'
      custom_rules:
        items:
          $ref: '#/definitions/model.CustomRule'
        type: array
      logs:
        $ref: '#/definitions/model.LogsSettings'
      privacy:
        $ref: '#/definitions/model.Privacy'
      profile_id:
        type: string
      security:
        $ref: '#/definitions/model.Security'
      statistics:
        $ref: '#/definitions/model.StatisticsSettings'
    required:
    - advanced
    - logs
    - privacy
    - profile_id
    - security
    - statistics
    type: object
  model.ProfileUpdate:
    properties:
      operation:
        enum:
        - remove
        - add
        - replace
        - move
        - copy
        type: string
      path:
        enum:
        - /name
        - /settings/statistics/enabled
        - /settings/logs/enabled
        - /settings/logs/log_clients_ips
        - /settings/logs/log_domains
        - /settings/logs/retention
        - /settings/privacy/default_rule
        - /settings/privacy/subdomains_rule
        - /settings/security/dnssec/enabled
        - /settings/security/dnssec/send_do_bit
        - /settings/advanced/recursor
        type: string
      value: {}
    required:
    - operation
    - path
    - value
    type: object
...
  model.Retention:
    enum:
    - 1h
    - 6h
    - 1d
    - 1w
    - 1m
    type: string
    x-enum-varnames:
    - RetentionOneHour
    - RetentionSixHours
    - RetentionOneDay
    - RetentionOneWeek
    - RetentionOneMonth
  model.Security:
    properties:
      dnssec:
        $ref: '#/definitions/model.DNSSECSettings'
    required:
    - dnssec
    type: object
  model.StatisticsAggregated:
    properties:
      total:
        description: 'Note: "total" needs to be the same as in the repository mongo
          query'
        type: integer
    type: object
  model.StatisticsSettings:
    properties:
      enabled:
        type: boolean
    required:
    - enabled
    type: object
...
  model.TOTPBackup:
    properties:
      backup_codes:
        items:
          type: string
        type: array
    type: object
  model.TOTPNew:
    properties:
      account:
        type: string
      secret:
        type: string
      uri:
        type: string
    type: object
  model.TotpSettings:
    properties:
      enabled:
        description: Indicates if TOTP is enabled.
        type: boolean
    type: object
...
  requests.ConfirmResetPasswordBody:
    properties:
      new_password:
        type: string
      otp:
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
...
  requests.ProfileUpdates:
    properties:
      updates:
        items:
          $ref: '#/definitions/model.ProfileUpdate'
        type: array
    required:
    - updates
    type: object
  requests.ResetPasswordBody:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  requests.TotpReq:
    properties:
      otp:
        maxLength: 8
        minLength: 6
        type: string
    required:
    - otp
    type: object
...
info:
  contact: {}
  description: modDNS REST API
  title: modDNS REST API
  version: "1.0"
paths:
...
  /api/v1/accounts/mfa/totp/disable:
    post:
      consumes:
      - application/json
      description: Disable TOTP
      parameters:
      - description: Disable TOTP request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.TotpReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Account'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      summary: Disable TOTP
      tags:
      - Account
  /api/v1/accounts/mfa/totp/enable:
    post:
      consumes:
      - application/json
      description: Enable TOTP
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TOTPNew'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      summary: Enable TOTP
      tags:
      - Account
  /api/v1/accounts/mfa/totp/enable/confirm:
    post:
      consumes:
      - application/json
      description: Confirm TOTP
      parameters:
      - description: Confirm TOTP request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.TotpReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TOTPBackup'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      summary: Confirm TOTP
      tags:
      - Account
  /api/v1/accounts/reset-password:
    post:
      consumes:
      - application/json
      description: Send reset password email
      parameters:
      - description: Send reset password email request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.ResetPasswordBody'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      summary: Send reset password email
      tags:
      - Account
...
  /api/v1/profiles:
    get:
      description: Get profiles data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Profile'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      security:
      - ApiKeyAuth: []
      summary: Get profiles data
      tags:
      - Profile
    post:
      consumes:
      - application/json
      description: Create profile
      parameters:
      - description: Create profile request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.createProfileBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Profile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      security:
      - ApiKeyAuth: []
      summary: Create profile
      tags:
      - Profile
  /api/v1/profiles/{id}:
    delete:
      description: Delete profile
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete profile
      tags:
      - Profile
    get:
      description: Get profile data
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Profile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      security:
      - ApiKeyAuth: []
      summary: Get profile data
      tags:
      - Profile
    patch:
      consumes:
      - application/json
      description: Update profile
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      - description: Update profile
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.ProfileUpdates'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Profile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      security:
      - ApiKeyAuth: []
      summary: Update profile
      tags:
      - Profile
...
  /api/v1/profiles/{id}/statistics:
    get:
      description: Get statistics data for a profile
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      - default: '"LAST_MONTH"'
        description: specify timespan for query
        in: query
        name: timespan
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.StatisticsAggregated'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      security:
      - ApiKeyAuth: []
      summary: Get statistics data for a profile
      tags:
      - Statistics
...
  /api/v1/verify/reset-password:
    post:
      consumes:
      - application/json
      description: Confirm password reset
      parameters:
      - description: Confirm password reset request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.ConfirmResetPasswordBody'
      - description: MFA OTP code
        in: header
        name: x-mfa-code
        type: string
      - collectionFormat: csv
        description: MFA methods
        in: header
        items:
          type: string
        name: x-mfa-methods
        type: array
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      summary: Confirm password reset
      tags:
      - Verification
...
swagger: "2.0"